Adding conventional `npm start` command to package.json
It's common node practice to have a start script, which is special to
npm because it allows you to run `npm start` which is the usual way to
run an application.

This also adds the config flag "private": true which prevents the
package from being published to npm by accident. This could prevent the
unlikely case that someone accidentally publishes their api keys